# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuantSignal - AI-powered quant signal validation platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

/**
 * User's market bias
 */
export type AnalysisRequestMarketBias =
  (typeof AnalysisRequestMarketBias)[keyof typeof AnalysisRequestMarketBias];

export const AnalysisRequestMarketBias = {
  bullish: "bullish",
  bearish: "bearish",
  neutral: "neutral",
} as const;

export interface AnalysisRequest {
  /** Trading symbol e.g. BTCUSDT, AAPL, NIFTY */
  symbol: string;
  /** Chart timeframe e.g. 5m, 15m, 1h, 1d */
  timeframe: string;
  /** Optional plain text strategy note */
  strategyNote?: string;
  /** Optional entry price */
  entryPrice?: number;
  /** Optional stop loss price */
  stopLoss?: number;
  /** Optional target price */
  target?: number;
  /** User's market bias */
  marketBias?: AnalysisRequestMarketBias;
  /** Optional uploaded screenshot file ID */
  screenshotFileId?: string;
  /** Optional uploaded trade CSV file ID */
  tradeFileId?: string;
}

export type SignalResultSignal =
  (typeof SignalResultSignal)[keyof typeof SignalResultSignal];

export const SignalResultSignal = {
  Buy: "Buy",
  Sell: "Sell",
  Hold: "Hold",
  Avoid: "Avoid",
} as const;

export type SignalResultMarketRegime =
  (typeof SignalResultMarketRegime)[keyof typeof SignalResultMarketRegime];

export const SignalResultMarketRegime = {
  trend: "trend",
  range: "range",
  breakout: "breakout",
  breakdown: "breakdown",
  volatile: "volatile",
} as const;

export type SignalResultFinalBanner =
  (typeof SignalResultFinalBanner)[keyof typeof SignalResultFinalBanner];

export const SignalResultFinalBanner = {
  Safe: "Safe",
  Caution: "Caution",
  Avoid: "Avoid",
} as const;

export interface ServiceBreakdown {
  ingestion: string;
  signal: string;
  risk: string;
  regime: string;
  judge: string;
}

export interface SignalResult {
  id: string;
  symbol: string;
  timeframe: string;
  signal: SignalResultSignal;
  /** 0-100 confidence score */
  confidenceScore: number;
  /** 0-100 risk score (higher = more risky) */
  riskScore: number;
  marketRegime: SignalResultMarketRegime;
  /** Plain English explanation */
  explanation: string;
  /** Price level or condition that invalidates the signal */
  invalidationZone: string;
  /** Suggested stop loss price */
  stopLossSuggestion?: number;
  finalBanner: SignalResultFinalBanner;
  serviceBreakdown: ServiceBreakdown;
  createdAt: string;
}

export interface UploadResult {
  fileId: string;
  filename: string;
  size: number;
  message: string;
}

export type TradeRecordSide =
  (typeof TradeRecordSide)[keyof typeof TradeRecordSide];

export const TradeRecordSide = {
  buy: "buy",
  sell: "sell",
  long: "long",
  short: "short",
} as const;

export interface TradeRecord {
  date: string;
  symbol: string;
  side: TradeRecordSide;
  entry: number;
  exit: number;
  pnl: number;
}

export interface TradeUploadResult {
  fileId: string;
  filename: string;
  totalRecords: number;
  parsedRecords: TradeRecord[];
  message: string;
}

export type SignalHistoryItemSignal =
  (typeof SignalHistoryItemSignal)[keyof typeof SignalHistoryItemSignal];

export const SignalHistoryItemSignal = {
  Buy: "Buy",
  Sell: "Sell",
  Hold: "Hold",
  Avoid: "Avoid",
} as const;

export type SignalHistoryItemMarketRegime =
  (typeof SignalHistoryItemMarketRegime)[keyof typeof SignalHistoryItemMarketRegime];

export const SignalHistoryItemMarketRegime = {
  trend: "trend",
  range: "range",
  breakout: "breakout",
  breakdown: "breakdown",
  volatile: "volatile",
} as const;

export type SignalHistoryItemFinalBanner =
  (typeof SignalHistoryItemFinalBanner)[keyof typeof SignalHistoryItemFinalBanner];

export const SignalHistoryItemFinalBanner = {
  Safe: "Safe",
  Caution: "Caution",
  Avoid: "Avoid",
} as const;

export interface SignalHistoryItem {
  id: string;
  symbol: string;
  timeframe: string;
  signal: SignalHistoryItemSignal;
  confidenceScore: number;
  riskScore: number;
  marketRegime: SignalHistoryItemMarketRegime;
  finalBanner: SignalHistoryItemFinalBanner;
  createdAt: string;
}

export interface SignalHistoryResponse {
  items: SignalHistoryItem[];
  total: number;
  limit: number;
  offset: number;
}

export interface SettingsStatus {
  geminiApiKeyConfigured: boolean;
  databaseConnected: boolean;
  version: string;
}

export type UploadScreenshotBody = {
  file: Blob;
  symbol?: string;
  timeframe?: string;
};

export type UploadTradesBody = {
  file: Blob;
};

export type GetSignalHistoryParams = {
  limit?: number;
  offset?: number;
};