# lib/api-zod/src/generated/types/signalResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuantSignal - AI-powered quant signal validation platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceBreakdown } from "./serviceBreakdown";
import type { SignalResultFinalBanner } from "./signalResultFinalBanner";
import type { SignalResultMarketRegime } from "./signalResultMarketRegime";
import type { SignalResultSignal } from "./signalResultSignal";

export interface SignalResult {
  id: string;
  symbol: string;
  timeframe: string;
  signal: SignalResultSignal;
  /** 0-100 confidence score */
  confidenceScore: number;
  /** 0-100 risk score (higher = more risky) */
  riskScore: number;
  marketRegime: SignalResultMarketRegime;
  /** Plain English explanation */
  explanation: string;
  /** Price level or condition that invalidates the signal */
  invalidationZone: string;
  /** Suggested stop loss price */
  stopLossSuggestion?: number;
  finalBanner: SignalResultFinalBanner;
  serviceBreakdown: ServiceBreakdown;
  createdAt: Date;
}