# lib/api-zod/src/generated/types/tradeUploadResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuantSignal - AI-powered quant signal validation platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TradeRecord } from "./tradeRecord";

export interface TradeUploadResult {
  fileId: string;
  filename: string;
  totalRecords: number;
  parsedRecords: TradeRecord[];
  message: string;
}