# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuantSignal - AI-powered quant signal validation platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResult {
  fileId: string;
  filename: string;
  size: number;
  message: string;
}